Made some progress
