+++ packages/core/package.json
@@ -68,7 +68,7 @@
     "lint:fix": "yarn lint:prettier --write && yarn lint:eslint --fix",
     "lint:eslint": "eslint './{src,test}/**/*.{ts,tsx}'",
     "lint:prettier": "yarn prettier './{src,test}/**/*.{ts,tsx}'",
-    "generate": "ts-node ./generate.ts"
+    "generate": "ts-node ./generate/generate.ts"
   },
   "publishConfig": {
     "access": "public"
