+++ packages/core/tsconfig.json
@@ -19,6 +19,7 @@
     "src",
     "src/**/*.json",
     "internal",
-    "build-typechain"
+    "build-typechain",
+    "build-typechain/**/*.json",
   ]
 }
